Add files via upload
Keylogger Software
Keylogger software is a good beginner-level cyber security project. A keylogger is software used to record every keystroke made by the user on their keyboard devices.

Note: Keystrokes are interactions that we make with the keyboard buttons. Pressing a single key in a physical or a virtual keyboard is an example of a keystroke.

The motive of this software is to monitor the user’s activity to spot any unusual or threatening activities on the devices. This project is used in industries to keep an eye on their employees to ensure that they abide by the company’s policies. If you’re looking forward to working in the internal security sector of a company, this project gives you the exposure that you need.

You can extend the project to virtual keyboards, clipping loggers, screen loggers, and activity trackers. Since keyloggers are used at the ground level of complex security monitoring software, it can be a good start for a beginner. 

Sample source code: GitHub

The technology used in the example: are Node.js and JavaScript.